api/requirements.txt
Django==2.1.2
gunicorn==19.9.0
psycopg2==2.7.6
django-cors-headers==2.4.0
newspaper3k==0.2.8
